requirements.txt
pypdf2
python-docx